isolate statement feature into a trait
